MOMO-28 feat: apply redux state (14)
튜토리얼의 BottomButton 을 common component 로 교체
